Add labels for local domain names with OrbStack
All OrbStack needs to access a container using a domain name (usually *.local) is to add a label to that service with the domain name you want to use. OrbStack handles provisioning an SSL certificate for that hostname. The other nice thing about it is that you don’t need to include a port. So, if you have an app server running on port 3000, instead of accessing it with **http://localhost:3000**, you can access it at **https://webapp.local**. 🙂

I’m sure that I didn’t cover all of the services that can be accessed individually. We can add lablels for anything I missed.